requirements.txt
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
tabulate>=0.9.0
python-dotenv>=1.0.0
web3>=6.15.0
# Ronin-specific helpers (opcional):
# eth-account>=0.10.0
